fix and test: add main.cc for test

+++ src/util/coding.cc
@@ -60,7 +60,7 @@ char* EncodeVarint64(char *dst, uint64_t value) {
     return reinterpret_cast<char*>(ptr);
 }
 
-void PutVarint64(std::string *dst, uint32_t value) {
+void PutVarint64(std::string *dst, uint64_t value) {
     char buf[10];
     char* ptr = EncodeVarint64(buf, value);
     dst->append(buf, ptr - buf);
